refactor: extract ruleGroupSchema to shared schemas/rule-group.ts
Remove duplicate ruleSchema/ruleGroupDepth*/ruleGroupSchema definitions from
stored-agents.ts, stored-prompt-blocks.ts, and prompt-block-versions.ts.
All 3 files now import from the shared module.
