+++ packages/server/src/server/schemas/prompt-block-versions.ts
@@ -1,4 +1,5 @@
 import z from 'zod';
+import { ruleGroupSchema } from './rule-group';
 import {
   listVersionsQuerySchema,
   compareVersionsQuerySchema,
@@ -31,40 +32,6 @@ export const promptBlockVersionIdPathParams = z.object({
 // ============================================================================
 // Response Schemas
 // ============================================================================
-
-const ruleSchema = z.object({
-  field: z.string(),
-  operator: z.enum([
-    'equals',
-    'not_equals',
-    'contains',
-    'not_contains',
-    'greater_than',
-    'less_than',
-    'greater_than_or_equal',
-    'less_than_or_equal',
-    'in',
-    'not_in',
-    'exists',
-    'not_exists',
-  ]),
-  value: z.unknown(),
-});
-
-const ruleGroupDepth2 = z.object({
-  operator: z.enum(['AND', 'OR']),
-  conditions: z.array(ruleSchema),
-});
-
-const ruleGroupDepth1 = z.object({
-  operator: z.enum(['AND', 'OR']),
-  conditions: z.array(z.union([ruleSchema, ruleGroupDepth2])),
-});
-
-const ruleGroupSchema = z.object({
-  operator: z.enum(['AND', 'OR']),
-  conditions: z.array(z.union([ruleSchema, ruleGroupDepth1])),
-});
 
 export const promptBlockVersionSchema = z.object({
   id: z.string().describe('Unique identifier for the version (UUID)'),
