+++ packages/server/src/server/schemas/stored-agents.ts
@@ -2,6 +2,7 @@ import z from 'zod';
 import { paginationInfoSchema, createPagePaginationSchema, statusQuerySchema } from './common';
 import { defaultOptionsSchema } from './default-options';
 import { serializedMemoryConfigSchema } from './memory-config';
+import { ruleGroupSchema } from './rule-group';
 
 // ============================================================================
 // Path Parameter Schemas
@@ -45,47 +46,6 @@ export const listStoredAgentsQuerySchema = createPagePaginationSchema(100).exten
 // ============================================================================
 // Body Parameter Schemas
 // ============================================================================
-
-/**
- * Rule and RuleGroup schemas for conditional prompt block evaluation.
- */
-const ruleSchema = z.object({
-  field: z.string(),
-  operator: z.enum([
-    'equals',
-    'not_equals',
-    'contains',
-    'not_contains',
-    'greater_than',
-    'less_than',
-    'greater_than_or_equal',
-    'less_than_or_equal',
-    'in',
-    'not_in',
-    'exists',
-    'not_exists',
-  ]),
-  value: z.unknown(),
-});
-
-/**
- * Rule group schema with a fixed nesting depth (3 levels) to avoid
- * infinite recursion when converting to JSON Schema / OpenAPI.
- */
-const ruleGroupDepth2 = z.object({
-  operator: z.enum(['AND', 'OR']),
-  conditions: z.array(ruleSchema),
-});
-
-const ruleGroupDepth1 = z.object({
-  operator: z.enum(['AND', 'OR']),
-  conditions: z.array(z.union([ruleSchema, ruleGroupDepth2])),
-});
-
-const ruleGroupSchema = z.object({
-  operator: z.enum(['AND', 'OR']),
-  conditions: z.array(z.union([ruleSchema, ruleGroupDepth1])),
-});
 
 /**
  * Scorer config schema with optional sampling and rules
