+++ packages/server/src/server/schemas/stored-prompt-blocks.ts
@@ -1,6 +1,7 @@
 import z from 'zod';
 
 import { paginationInfoSchema, createPagePaginationSchema, statusQuerySchema } from './common';
+import { ruleGroupSchema } from './rule-group';
 
 // ============================================================================
 // Path Parameter Schemas
@@ -35,47 +36,6 @@ export const listStoredPromptBlocksQuerySchema = createPagePaginationSchema(100)
 // ============================================================================
 // Body Parameter Schemas
 // ============================================================================
-
-/**
- * Rule and RuleGroup schemas for conditional prompt block evaluation.
- */
-const ruleSchema = z.object({
-  field: z.string(),
-  operator: z.enum([
-    'equals',
-    'not_equals',
-    'contains',
-    'not_contains',
-    'greater_than',
-    'less_than',
-    'greater_than_or_equal',
-    'less_than_or_equal',
-    'in',
-    'not_in',
-    'exists',
-    'not_exists',
-  ]),
-  value: z.unknown(),
-});
-
-/**
- * Rule group schema with a fixed nesting depth (3 levels) to avoid
- * infinite recursion when converting to JSON Schema / OpenAPI.
- */
-const ruleGroupDepth2 = z.object({
-  operator: z.enum(['AND', 'OR']),
-  conditions: z.array(ruleSchema),
-});
-
-const ruleGroupDepth1 = z.object({
-  operator: z.enum(['AND', 'OR']),
-  conditions: z.array(z.union([ruleSchema, ruleGroupDepth2])),
-});
-
-const ruleGroupSchema = z.object({
-  operator: z.enum(['AND', 'OR']),
-  conditions: z.array(z.union([ruleSchema, ruleGroupDepth1])),
-});
 
 const snapshotConfigSchema = z.object({
   name: z.string().describe('Display name of the prompt block'),
